# Queue<T> optimization of (Try)Dequeue