# Description

* enabled RCE on array-access
* in MoveNext the JIT can produce better code (see https://github.com/dotnet/corefx/pull/25970#discussion_r157375261)

By _Dequeue_ the effect on value types is not so big, than for reference types (two array accesses).

This PR is a kind of extension to #17318, similar to https://github.com/dotnet/corefx/pull/26086

# Benchmarks

## Notes

Code for benchmarks lives [here](https://github.com/gfoidl/Benchmarks/tree/master/corefx/System/Collections/Generic/Queue/source/QueueBenchmarks)

Due the use of http://benchmarkdotnet.org the benchmarks were done a couple of times, because some crazy results with perf x2 were reported and this seems too strange. The results shown here are the more realistic ones. Individual results are in the linked repo above.
The changes from this PR never showed a decrease in perf.

## Dequeue

``` ini

BenchmarkDotNet=v0.10.11, OS=ubuntu 16.04
Processor=Intel Xeon CPU 2.60GHz, ProcessorCount=2
.NET Core SDK=2.1.3
  [Host]     : .NET Core 2.0.4 (Framework 4.6.0.0), 64bit RyuJIT
  DefaultJob : .NET Core 2.0.4 (Framework 4.6.0.0), 64bit RyuJIT


```
|          Method |      Mean |     Error |    StdDev | Scaled | ScaledSD |
|---------------- |----------:|----------:|----------:|-------:|---------:|
| Dequeue_Default | 0.9702 ns | 0.1013 ns | 0.1421 ns |   1.00 |     0.00 |
|     Dequeue_New | 0.8880 ns | 0.1042 ns | 0.1770 ns |   0.93 |     0.22 |

## TryDequeue

``` ini

BenchmarkDotNet=v0.10.11, OS=ubuntu 16.04
Processor=Intel Xeon CPU 2.60GHz, ProcessorCount=2
.NET Core SDK=2.1.3
  [Host]     : .NET Core 2.0.4 (Framework 4.6.0.0), 64bit RyuJIT
  DefaultJob : .NET Core 2.0.4 (Framework 4.6.0.0), 64bit RyuJIT


```
|             Method |     Mean |     Error |    StdDev | Scaled | ScaledSD |
|------------------- |---------:|----------:|----------:|-------:|---------:|
| TryDequeue_Default | 3.990 ns | 0.1666 ns | 0.2545 ns |   1.00 |     0.00 |
|     TryDequeue_New | 3.559 ns | 0.1585 ns | 0.2273 ns |   0.90 |     0.08 |

# Notes

## Enqueue

Didn't find a way to improve perf, besides the effect of better JIT-codegen in _MoveNext_.
Remained Untouched.

## (Try)Peek

Didn't find a way to improve perf, and "RCE-if" does not bring a win, it's getting rather slower, because the code gets bigger. I didn't keep benchmark-records for this case.
Remained untouched.