@karelz 2018-01-27
@safern what are next steps of this PR? Do we take it, or not?